refactor : utils file 정리
