+++ apps/web/containers/Notices/Notice/index.tsx
@@ -7,8 +7,7 @@ import { useEffect, useState } from 'react';
 import { PATHS } from '@/constants';
 import { apiBe, getHeaders } from '@/services';
 import { DecodedToken, INotice } from '@/types';
-import { decodeToken_csr } from '@/utils';
-import { dateSplit } from '@/utils/dateTime';
+import { decodeToken_csr, dateSplit } from '@/utils';
 
 import styles from './index.module.scss';
 import NoticeRegister from '../NoticeRegister';
